pull ScopedView out of view_builder.rs
