+++ presenter/src/view_builder.rs
@@ -1,5 +1,7 @@
 use crate::input_processor::Subscriber;
-use crate::{Context, ContextPath, ContextScope, InputProcessor, ProcessorRecord, Support, View};
+use crate::{
+    Context, ContextPath, ContextScope, InputProcessor, ProcessorRecord, ScopedView, Support, View,
+};
 use emergent_drawing::{
     Bounds, DrawingBounds, DrawingFastBounds, MeasureText, Rect, Text, Transform, Transformed,
 };
@@ -118,44 +120,6 @@ impl<Msg> ViewBuilder<Msg> {
                 .map(|s| (ContextPath::new(), s))
                 .collect(),
         )
-    }
-}
-
-/// This type allows a parent view building function to apply certain mutations to the
-/// scoped view.
-pub struct ScopedView<Msg> {
-    view: View<Msg>,
-}
-
-impl<Msg> ScopedView<Msg> {
-    pub(crate) fn new(view: View<Msg>) -> Self {
-        Self { view }
-    }
-
-    fn map(self, f: impl Fn(View<Msg>) -> View<Msg>) -> Self {
-        Self { view: f(self.view) }
-    }
-
-    pub fn fast_bounds(&self, measure: &dyn MeasureText) -> DrawingBounds {
-        self.view.fast_bounds(measure)
-    }
-
-    pub fn transformed(self, transform: impl Into<Transform>) -> Self {
-        let transform = transform.into();
-        self.map(move |v| v.transformed(transform.clone()))
-    }
-
-    pub fn in_area(self) -> Self {
-        self.map(|v| v.in_area())
-    }
-
-    pub fn presentation_scoped(self, scope: impl Into<PresentationScope>) -> Self {
-        let scope = scope.into();
-        self.map(|v| v.presentation_scoped(scope.clone()))
-    }
-
-    pub(crate) fn into_view(self) -> View<Msg> {
-        self.view
     }
 }
